+++ CHANGELOG.md
@@ -4,10 +4,10 @@ Changelog
 ## V2.0.0
 
 - Removed built-in `Promise` polyfill
-- Some APIs that returned Promises now return bare values
+- Some APIs that returned `Promise`s now return bare values.
 - Removed Bower support
 - Added ES module support
-- Added `arrayBufferToBinaryString` and `binaryStringToArrayBuffer`
+- Added `arrayBufferToBinaryString()` and `binaryStringToArrayBuffer()`
 
 ## v1.0.0